Apply patch for HARMONY-869: [classlib][logging] run of logging test lead to remove the deploy\jdk\jre\lib\logging.properties file
git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@422304 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/logging/src/test/java/org/apache/harmony/logging/tests/java/util/logging/util/DefaultPropertyHelper.java
@@ -88,9 +88,11 @@ public class DefaultPropertyHelper {
 	}
 
 	public synchronized static void reset(File bak) {
-		File file = new File(propertyPath);
-		file.delete();
-		bak.renameTo(file);
+		if (bak.exists()) {
+			File file = new File(propertyPath);
+			file.delete();
+			bak.renameTo(file);
+		}
 	}
 
 }
